# Use TimerOutputs.jl to measure time and memory performance