--- a/REQUIRE
+++ b/REQUIRE
@@ -5,3 +5,4 @@ HDF5 0.7
 DataFrames
 Formatting
 Logging
+TimerOutputs