--- a/test/test_elasticity_2d_linear_with_surface_load.jl
+++ b/test/test_elasticity_2d_linear_with_surface_load.jl
@@ -55,3 +55,5 @@ using JuliaFEM.Testing
     u3_expected = f/E*[-nu, 1] + g/(2*E)*[-nu, 1]
     @test isapprox(u3, u3_expected)
 end
+
+print_statistics()