- See issue #98.
- Added `TimerOutputs` to `REQUIRE` file.
- JuliaFEM has now `TimerOutput()` handle `to` and function `print_statistics()` to get summary of time and memory usage.
- `TimerOutputs.jl` is used to measure performance of `LinearSolver` at the moment and tested in file `test_elasticity_2d_linear_with_surface_load.jl`

Example output:

```
 ──────────────────────────────────────────────────────────────────────────────
                                       Time                   Allocations
                               ──────────────────────   ───────────────────────
       Tot / % measured:            14.6s / 33.7%            623MiB / 29.8%

 Section               ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────
 assemble problems          1    3.19s  65.0%   3.19s    146MiB  78.8%   146MiB
 solve linear system        1    1.38s  28.1%   1.38s   30.6MiB  16.5%  30.6MiB
 update problems            1    177ms  3.61%   177ms   5.81MiB  3.13%  5.81MiB
 initialize solver          1    162ms  3.30%   162ms   2.85MiB  1.54%  2.85MiB
 ──────────────────────────────────────────────────────────────────────────────
```